backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-compress==1.15  # gzip压缩支持，加速数据传输
psycopg2-binary==2.9.9
python-dotenv==1.0.1
werkzeug==3.0.4


